Module reactive_graph_graphql_schema::mutation::instances::relation_instance_id

Structs

GraphQLRelationInstanceId

The primary key of an edge consists of the outbound id, the type name and the inbound id.